scaffold del servidor
