+++ backend/.env.example
@@ -1,0 +1,9 @@
+PORT=3001
+PASS_TWITCH=oauth:bxbitggxf9q0s45vvkg83a9wte84kv
+
+
+CLIENT_ID_TWITCH = cuozlwcgreix8bpowwg9h6dxdcfhgo
+CLIENT_SECRET_TWITCH = ikxd3p6hrm662r6489ciiw0z60ps20
+ACCESS_TOKEN = 672306e4y8g07u0qo52zm2z0bqlz7y
+
+
