--- backend/.env.local
@@ -1,6 +1,0 @@
-VITE_CLIENT_ID_TWITCH = cuozlwcgreix8bpowwg9h6dxdcfhgo
-VITE_CLIENT_SECRET_TWITCH = ikxd3p6hrm662r6489ciiw0z60ps20
-VITE_PASS_TWITCH = oauth:bxbitggxf9q0s45vvkg83a9wte84kv
-VITE_ACCESS_TOKEN = 672306e4y8g07u0qo52zm2z0bqlz7y
-
-
