+++ backend/app.mjs
@@ -1,0 +1,25 @@
+import express from 'express'
+import * as dotenv from 'dotenv'
+import cors from 'cors'
+import http from 'node:http'
+import { Server } from 'socket.io'
+import { twitch } from './config/twitch.mjs'
+
+dotenv.config()
+
+/* Configuramos el puerto */
+const PORT = process.env.PORT || 3000
+
+/* Configuramos el servidor de express */
+const app = express()
+app.use(cors()) // hacemos uso de las cors, para evitar el error de origen cruzado en el navegador
+
+const httpServer = http.createServer(app)
+const io = new Server(httpServer)
+
+twitch({ io })
+
+// Iniciar el servidor y escuchar solicitudes
+httpServer.listen(PORT, () => {
+  console.log(`Servidor iniciado en http://localhost:${PORT}`)
+})
