+++ backend/package.json
@@ -11,8 +11,12 @@
   "license": "ISC",
   "dependencies": {
     "cors": "2.8.5",
+    "dotenv": "^16.0.3",
     "express": "^4.18.2",
     "socket.io": "^4.6.1",
     "tmi.js": "1.8.5"
+  },
+  "eslintConfig": {
+    "extends": "./node_modules/standard/eslintrc.json"
   }
 }
