+++ frontend/src/styles/Chip.module.css
@@ -1,6 +1,6 @@
 .chip {
   background-color: #38bdf833;
-  padding: .2rem .8rem;
+  padding: .5rem .8rem;
   color: #38BDF8;
-  border-radius: 1rem;
+  border-radius: .3em;
 }
